v3: fixes #7
Fix #7: There's now a version of /data with url & name parameters, name as an alternative to filename. This allows you to send the name shown in Tableau Public desktop for the data source, instead of the internal .hyper filename.. This is important because I've seen those .hyper filenames/paths change multiple times across publishes. This way you won't need to worry about any of that, and should be set (until the publisher renames the data source!)

There's also /datasources?url which returns all names and corresponding filenames. In the off chance there's a data source that has an extract, but no name, it'll just repeat the filename. I haven't seen a .twbx file like that yet.

For now the old endpoints are faster.